New dependencies, project structure and code
Added backend code and changed paths in html.
