+++ app.js
@@ -1,49 +1,16 @@
-const button = document.getElementById('submit-button')
-button.onclick = sendImg
+const express = require('express')
+const path = require('path')
 
-function sendImg () {
-  /* TODO:
-  - make error no link (toast message from Bootstrap)
-  */
-  const imgUrl = document.getElementById('image-url').value
-  console.log('imgUrl: ' + Boolean(imgUrl))
+const app = express()
 
-  if (!imgUrl) { // if there is no link
-    throw new Error('No URL passed')
-  } else if (imgUrl.match('^(ftp|http|https)://[^ "]+$')) {
-    const options = {
-      method: 'POST',
-      headers: {
-        'content-type': 'application/json',
-        'X-RapidAPI-Host': 'image-labeling1.p.rapidapi.com',
-        'X-RapidAPI-Key': '' // paste here api key secretly
-      },
-      body: `{'url':'${imgUrl}'}`
-    }
+app.use(express.static(path.join(__dirname, 'node_modules/bootstrap/dist/')))
+app.use(express.static(path.join(__dirname, 'views')))
+app.use(express.static(path.join(__dirname, 'assets')))
 
-    console.log('Sending imgUrl and getting response...')
-    fetch('https://image-labeling1.p.rapidapi.com/img/label', options) // getting response
-      .then(response => response.json())
-      .then(processResponse)
-      .catch(err => console.error(err))
-  }
-}
+app.get('/', (req, res) => {
+  res.sendFile(path.join(__dirname, 'views/index.html'))
+})
 
-function processResponse (response) {
-  const imageInfoDiv = document.getElementById('image-info')
-  const infoHeader = imageInfoDiv.children[0]
-
-  console.log(response)
-
-  while (imageInfoDiv.childNodes.length > 5) { // counted nodes by myself NOT SCALABLE!!!
-    imageInfoDiv.removeChild(imageInfoDiv.lastChild)
-  }
-  infoHeader.textContent = 'Info'
-  for (const label in response) { // enumerating (iterating through indexes)
-    // take only two numbers after point into string
-    const strmatch = response[label].toString().slice(2, 4)
-    const labelPara = document.createElement('p')
-    labelPara.textContent = label + ': ' + strmatch + '%'
-    imageInfoDiv.appendChild(labelPara)
-  }
-}
+app.listen(5000, () => {
+  console.log('Listening on port ' + 5000)
+})
