+++ assets/favicon/android-chrome-512x512.png
